feat(tasks): create tasks folder to store Gulp tasks
