+++ package.json
@@ -19,6 +19,7 @@
     "gulp-imagemin": "^3.0.1",
     "merge-stream": "^1.0.0",
     "plylog": "^0.4.0",
-    "polymer-build": "^0.1.0"
+    "polymer-build": "^0.1.0",
+    "require-dir": "^0.3.0"
   }
 }
